+++ src/test/java/org/chocosolver/solver/learn/LazyImplicationsTest.java
@@ -3,8 +3,8 @@
  *
  * Copyright (c) 2018, IMT Atlantique. All rights reserved.
  *
- * Licensed under the BSD 4-clause license. See LICENSE file in the project root for full license
- * information.
+ * Licensed under the BSD 4-clause license.
+ * See LICENSE file in the project root for full license information.
  */
 package org.chocosolver.solver.learn;